+++ js/data.js
@@ -1,0 +1,54 @@
+import { getRandomInteger } from './utils.js';
+
+const DESCRIPTION_NUMBER = 25;
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Артём',
+  'Алексей',
+  'Женя',
+  'Вова',
+  'Паша',
+  'Ирина',
+  'Аня',
+  'Наталья',
+];
+
+const createIdGenerator = (min, max) => {
+  let usedIds = [];
+  let ids = [];
+  for (let i = min; i <= max; i++) {
+    ids = [...ids, i];
+  }
+  return () => {
+    const notUsedIds = ids.filter((id) => !usedIds.includes(id));
+    const randomId = notUsedIds[getRandomInteger(0, notUsedIds.length - 1)];
+    usedIds = [...usedIds, randomId];
+    return randomId;
+  };
+};
+
+const getRandomArrayElement = (elements) =>
+  elements[getRandomInteger(0, elements.length - 1)];
+
+const generateId = createIdGenerator(1, 25);
+
+const createDescription = () => ({
+  id: generateId(),
+  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
+  message: getRandomArrayElement(COMMENTS),
+  name: getRandomArrayElement(NAMES),
+});
+
+const generateDescription = () =>
+  Array.from({ length: DESCRIPTION_NUMBER }, createDescription);
+
+export { generateDescription };
